last changes

+++ src/util/zs_path.cpp
@@ -7,14 +7,14 @@
 namespace zetscript{
 	namespace zs_path{
 		zs_string  get_directory(const zs_string & _path) {
-			size_t pos = _path.find_last_of("\\/");
+			int pos = _path.find_last_of("\\/");
 			 return (zs_string::npos == pos)
 				 ? ""
 				 : _path.substr(0, pos);
 		}
 
 		zs_string  get_filename(const zs_string & _path) {
-		  size_t found;
+		  int found;
 		  zs_string ss=_path;
 		  found=_path.find_last_of("/\\");
 		  if(found != zs_string::npos){
@@ -29,7 +29,7 @@ namespace zetscript{
 
 
 			zs_string fName(file);
-				size_t pos = fName.find_last_of(".");
+				int pos = fName.find_last_of(".");
 				if(pos == zs_string::npos)  //No extension.
 					return fName;
